# pao-back-end-development/docs/swagger.yaml
basePath: /
definitions:
  apierrors.APIErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierrors.AppError'
      message:
        description: Descriptive message providing details about the error.
        type: string
      status_code:
        description: HTTP status code representing the error type (e.g., 404 for Not
          Found).
        type: integer
      success:
        description: Success flag to indicate if the operation was successful.
        type: boolean
    type: object
  apierrors.AppError:
    properties:
      code:
        type: string
      field_errors:
        items:
          $ref: '#/definitions/apierrors.FieldError'
        type: array
      id:
        type: string
      message:
        type: string
    type: object
  apierrors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      value: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: support_cept@indiapost.gov.in
    name: API Support Team
    url: http://cept.gov.in/support
  description: A comprehensive API for budget related activities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://cept.gov.in/terms
  title: Address Service API
  version: "1.0"
paths:
...
  /v1/objection-file/download:
    get:
      consumes:
      - application/json
      description: Fetches a file from the specified storage bucket using the file
        key. The file is returned with the appropriate content type and can be viewed
        inline or downloaded.
      parameters:
      - description: File key
        in: query
        name: key
        required: true
        type: string
      - description: Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File retrieved successfully
          schema:
            type: file
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Retrieve a file from the server
      tags:
      - FILE
  /v1/objection-file/upload:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a file to the specified folder path on the server. The
        file must be of type JPG or PDF and should not exceed 1MB in size.
      parameters:
      - description: Destination folder path
        in: formData
        name: folderPath
        required: true
        type: string
      - description: File to upload (PDF OR JPG OR JPEG ONLY)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: File uploaded successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierrors.APIErrorResponse'
      summary: Upload a file to the server
      tags:
      - FILE
...
schemes:
- http
- https
swagger: "2.0"